enhance metadata panel behavior
relates https://github.com/MTES-MCT/monitorenv/issues/693
