+++ frontend/src/features/layersSelector/search/ResultsList/index.tsx
@@ -1,3 +1,4 @@
+import { closeMetadataPanel } from '@features/layersSelector/metadataPanel/slice'
 import { Checkbox } from '@mtes-mct/monitor-ui'
 import _ from 'lodash'
 import styled from 'styled-components'
@@ -41,6 +42,7 @@ export function ResultList({ searchedText }: ResultListProps) {
   const toggleRegulatory = () => {
     if (!isRegulatorySearchResultsVisible) {
       dispatch(setIsRegulatorySearchResultsVisible(true))
+      dispatch(closeMetadataPanel())
     }
     dispatch(setIsRegulatorySearchResultsExpanded(!isRegulatorySearchResultsExpanded))
     dispatch(setIsAmpSearchResultsExpanded(false))
@@ -48,16 +50,19 @@ export function ResultList({ searchedText }: ResultListProps) {
   const toggleAMPs = () => {
     if (!isAmpSearchResultsVisible) {
       dispatch(setIsAmpSearchResultsVisible(true))
+      dispatch(closeMetadataPanel())
     }
     dispatch(setIsRegulatorySearchResultsExpanded(false))
     dispatch(setIsAmpSearchResultsExpanded(!isAmpSearchResultsExpanded))
   }
 
   const toggleAMPVisibility = () => {
+    dispatch(closeMetadataPanel())
     dispatch(setIsAmpSearchResultsExpanded(false))
     dispatch(setIsAmpSearchResultsVisible(!isAmpSearchResultsVisible))
   }
   const toggleRegulatoryVisibility = () => {
+    dispatch(closeMetadataPanel())
     dispatch(setIsRegulatorySearchResultsExpanded(false))
     dispatch(setIsRegulatorySearchResultsVisible(!isRegulatorySearchResultsVisible))
   }
